ti-linux-kernel: Add patch to make Docker work
Docker requires a bunch of configs to be enabled. A list of these
options can be generated using the script at [1]. Many of the "optional"
configs aren't enabled, as a result of which Docker isn't able to pull
and run certain images. So add a patch that enables these configs.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>